chore: rename sse-go-pusher to go-services
