+++ go-services/go.mod
@@ -1,17 +1,15 @@
-module sse-go-pusher
+module go-services
 
 go 1.26.4
 
 require (
 	github.com/gin-gonic/gin v1.12.0
-	github.com/go-redis/redismock/v9 v9.2.0
 	github.com/jackc/pgx/v5 v5.10.0
 	github.com/joho/godotenv v1.5.1
 	github.com/redis/go-redis/v9 v9.21.0
 )
 
 require (
-	github.com/DATA-DOG/go-sqlmock v1.5.2 // indirect
 	github.com/bytedance/gopkg v0.1.3 // indirect
 	github.com/bytedance/sonic v1.15.0 // indirect
 	github.com/bytedance/sonic/loader v0.5.0 // indirect
