+++ go-services/internal/app/delivery/sse_request.go
