+++ go-services/internal/app/repositories/postgres/notification.go
@@ -3,7 +3,7 @@ package postgres
 import (
 	"context"
 
-	"sse-go-pusher/internal/domain"
+	"go-services/internal/domain"
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
